using better session management
